# src/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Search API",
    "description": "Privacy-respecting metasearch engine API",
    "version": "1.0.0",
    "contact": {
      "name": "Search Support"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "/api/v1",
      "description": "API v1"
    }
  ],
...
  "paths": {
...
    "/search": {
      "get": {
        "summary": "Search",
        "description": "Perform a search query across multiple engines",
        "operationId": "search",
        "tags": ["Search"],
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": true,
            "description": "Search query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "category",
            "in": "query",
            "description": "Search category",
            "schema": {
              "type": "string",
              "enum": ["general", "images", "videos", "news", "maps"],
              "default": "general"
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "Page number",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Results per page",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            }
          },
          {
            "name": "safe_search",
            "in": "query",
            "description": "Safe search level",
            "schema": {
              "type": "string",
              "enum": ["off", "moderate", "strict"],
              "default": "moderate"
            }
          },
          {
            "name": "lang",
            "in": "query",
            "description": "Language preference",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "engines",
            "in": "query",
            "description": "Comma-separated list of engines to use",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Search results",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Search (POST)",
        "description": "Perform a search query with JSON body",
        "operationId": "searchPost",
        "tags": ["Search"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SearchRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Search results",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/autocomplete": {
      "get": {
        "summary": "Autocomplete suggestions",
        "description": "Get autocomplete suggestions for a partial query",
        "operationId": "autocomplete",
        "tags": ["Search"],
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": true,
            "description": "Partial search query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Autocomplete suggestions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AutocompleteResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "API token authentication. Use token with prefix: usr_ for user tokens."
      },
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "session",
        "description": "Session cookie authentication"
      }
    },
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean",
            "example": false
          },
          "error": {
            "type": "string",
            "description": "Error code"
          },
          "message": {
            "type": "string",
            "description": "Human-readable error message"
          },
          "meta": {
            "$ref": "#/components/schemas/APIMeta"
          }
        }
      },
      "APIMeta": {
        "type": "object",
        "properties": {
          "request_id": {
            "type": "string"
          },
          "process_time_ms": {
            "type": "number"
          },
          "version": {
            "type": "string"
          }
        }
      },
...
      "SearchRequest": {
        "type": "object",
        "required": ["query"],
        "properties": {
          "query": {
            "type": "string"
          },
          "category": {
            "type": "string",
            "enum": ["general", "images", "videos", "news", "maps"],
            "default": "general"
          },
          "page": {
            "type": "integer",
            "minimum": 1,
            "default": 1
          },
          "limit": {
            "type": "integer",
            "minimum": 1,
            "maximum": 100,
            "default": 20
          },
          "engines": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "safe_search": {
            "type": "string",
            "enum": ["off", "moderate", "strict"]
          },
          "time_range": {
            "type": "string"
          },
          "language": {
            "type": "string"
          }
        }
      },
      "SearchResponse": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean"
          },
          "data": {
            "type": "object",
            "properties": {
              "query": {
                "type": "string"
              },
              "category": {
                "type": "string"
              },
              "results": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/SearchResult"
                }
              },
              "total_results": {
                "type": "integer"
              },
              "page": {
                "type": "integer"
              },
              "limit": {
                "type": "integer"
              },
              "has_more": {
                "type": "boolean"
              },
              "search_time_ms": {
                "type": "number"
              },
              "engines_used": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            }
          },
          "meta": {
            "$ref": "#/components/schemas/APIMeta"
          }
        }
      },
      "SearchResult": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "url": {
            "type": "string",
            "format": "uri"
          },
          "description": {
            "type": "string"
          },
          "engine": {
            "type": "string"
          },
          "score": {
            "type": "number"
          },
          "category": {
            "type": "string"
          },
          "thumbnail": {
            "type": "string",
            "format": "uri"
          },
          "date": {
            "type": "string"
          },
          "domain": {
            "type": "string"
          }
        }
      },
...
      "AutocompleteResponse": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean"
          },
          "data": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
    }
  }
}